requirements.txt
# FastAPI
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Pydantic
pydantic==2.6.1
pydantic-settings==2.2.1

# Supabase
supabase==2.28.0

# LangChain / LangGraph
langchain==0.1.9
langgraph==0.0.28
langchain-openai==0.0.8
langsmith==0.1.10

# OpenAI
openai==1.12.0

# Google APIs
google-api-python-client==2.118.0
google-auth==2.28.0
google-auth-oauthlib==1.2.0

# HTTP Client
httpx>=0.24,<0.26

# Email
resend==0.8.0

# Redis / Celery
celery==5.3.6
redis==5.0.1

# Utils
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0

# PDF Generation
weasyprint==61.0
